backend/requirements.txt
fastapi
uvicorn
python-multipart
tensorflow
Pillow
requests
pydantic
xgboost
